strlcat

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: aarchtou <aarchtou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/09 18:24:59 by aarchtou          #+#    #+#             */
-/*   Updated: 2022/10/22 16:54:17 by aarchtou         ###   ########.fr       */
+/*   Updated: 2022/10/22 18:13:41 by aarchtou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 	if (dst == 0 && dstsize == 0)
 		return (ft_strlen(src));
 	c = ft_strlen(dst);
-	if (dstsize <= (size_t)ft_strlen(dst))
+	if (dstsize <= ft_strlen(dst))
 		return (dstsize + ft_strlen(src));
 	while (src[d] != '\0' && c + 1 < dstsize)
 	{
